// mobile/src/types.ts
export enum DrinkCategory {
  Coffee = "咖啡",
  Matcha = "抹茶",
  Tea = "传统茶",
  FruitTea = "水果茶",
  PourOver = "手冲",
  MilkTea = "奶茶"
}